+++ ogn/src/Shared/core/core/math/Point2.cpp
@@ -21,8 +21,8 @@
 
 float32 Point2::distance(const Point2& other) const
 {
-	Vector2 v0(x, y);
-	Vector2 v1(other.x, other.y);
+	Vector2 v0((float32)x, (float32)y);
+	Vector2 v1((float32)other.x, (float32)other.y);
 	return v0.distance(v1);
 }
